Make room for media types of the future. Extend the ifmedia word to 64 bits. This changes numbers of the SIOCSIFMEDIA and SIOCGIFMEDIA ioctls and grows struct ifmediareq.
Old ifconfig and dhclient binaries can still assign addresses, however
the 'media' subcommand stops working. Recompiling ifconfig and dhclient
with new headers before a reboot should not be necessary unless in very
special circumstances where non-default media settings must be used to
get link and console access is not available.

There may be some MD fallout but that will be cleared up later.

ok deraadt miod
with help and suggestions from several sharks attending l2k15

+++ sys/dev/ic/if_wi.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: if_wi.c,v 1.161 2015/05/19 11:34:30 mpi Exp $	*/
+/*	$OpenBSD: if_wi.c,v 1.162 2015/06/24 09:40:54 mpi Exp $	*/
 
 /*
  * Copyright (c) 1997, 1998, 1999
@@ -2672,9 +2672,9 @@ wi_get_id(struct wi_softc *sc)
 STATIC int
 wi_sync_media(struct wi_softc *sc, int ptype, int txrate)
 {
-	int media = sc->sc_media.ifm_cur->ifm_media;
-	int options = IFM_OPTIONS(media);
-	int subtype;
+	uint64_t media = sc->sc_media.ifm_cur->ifm_media;
+	uint64_t options = IFM_OPTIONS(media);
+	uint64_t subtype;
 
 	switch (txrate) {
 	case 1:
